requirements.txt
streamlit==1.55.0
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-chroma==1.1.0
langchain-groq==1.1.2
langchain-huggingface==1.2.1
chromadb==1.5.5
sentence-transformers==5.3.0
python-dotenv==1.2.2
unstructured==0.18.32
